requirements.txt
flask
nltk
scikit-learn
numpy
azure-ai-textanalytics
python-dotenv